Added servermode commands, Improved permission handling, Added combat logger stats support, Fixed class bugs

+++ src/main/java/gg/revival/factions/core/deathbans/callbacks/DeathbanDurationCallback.java
@@ -2,6 +2,6 @@ package gg.revival.factions.core.deathbans.callbacks;
 
 public interface DeathbanDurationCallback {
 
-    void onLookupComplete(int duration);
+    void onLookupComplete(long duration);
 
 }
